+++ server/src/main/java/org/opensearch/index/fielddata/fieldcomparator/UnsignedLongValuesComparatorSource.java
@@ -10,7 +10,6 @@ package org.opensearch.index.fielddata.fieldcomparator;
 
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.index.NumericDocValues;
-import org.apache.lucene.index.SortedNumericDocValues;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.search.FieldComparator;
 import org.apache.lucene.search.LeafFieldComparator;
@@ -24,7 +23,6 @@ import org.opensearch.index.fielddata.FieldData;
 import org.opensearch.index.fielddata.IndexFieldData;
 import org.opensearch.index.fielddata.IndexNumericFieldData;
 import org.opensearch.index.fielddata.LeafNumericFieldData;
-import org.opensearch.index.fielddata.SingletonSortedNumericUnsignedLongValues;
 import org.opensearch.index.fielddata.SortedNumericUnsignedLongValues;
 import org.opensearch.index.search.comparators.UnsignedLongComparator;
 import org.opensearch.search.DocValueFormat;
@@ -61,8 +59,7 @@ public class UnsignedLongValuesComparatorSource extends IndexFieldData.XFieldCom
 
     private SortedNumericUnsignedLongValues loadDocValues(LeafReaderContext context) {
         final LeafNumericFieldData data = indexFieldData.load(context);
-        SortedNumericDocValues values = data.getLongValues();
-        return new SingletonSortedNumericUnsignedLongValues(values);
+        return FieldData.singleton(data.getLongValues());
     }
 
     private NumericDocValues getNumericDocValues(LeafReaderContext context, BigInteger missingValue) throws IOException {